issue_20 - refactoring and optimization
