--- src/main/java/com/yourrents/api/config/PropertyTypeConverter.java
@@ -1,41 +1,0 @@
-package com.yourrents.api.config;
-
-/*-
- * #%L
- * YourRents API
- * %%
- * Copyright (C) 2023 - 2024 Your Rents Team
- * %%
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * #L%
- */
-
-import com.yourrents.api.property.PropertyType;
-import org.jooq.impl.EnumConverter;
-
-public class PropertyTypeConverter extends EnumConverter<Object, PropertyType> {
-
-  public PropertyTypeConverter() {
-    super(Object.class, PropertyType.class);
-  }
-
-  @Override
-  public Object to(PropertyType propertyType) {
-    return propertyType.getTypeName();
-  }
-
-  @Override
-  public String toString() {
-    return super.toString();
-  }
-}
